--- a/src/streamable/mod.rs
+++ b/src/streamable/mod.rs
@@ -1,5 +1,4 @@
 pub mod bytes;
-// pub mod coin;
 // pub mod fullblock;
 // pub mod program;
 pub mod de;