--- a/wheel/src/lib.rs
+++ b/wheel/src/lib.rs
@@ -1,5 +1,6 @@
 mod adapt_response;
 mod api;
+mod coin;
 mod from_json_dict;
 mod run_generator;
 mod run_program;