This patch implements `Coin` in rust. Apart from supporting "Streamable" serialization/deserialization, from_json, to_json etc. `Coin` has one function: "name()", which computes the coin name.

Using a rust implementation for `Coin` provides two main improvements:

1. The `name()` function can be implemented efficiently (without allocating any any heap space)
2. The data structure of the coin is compact and more efficient to access because of that (all fields are fixed size)

My benchmarks suggests that this shaves off about 22% CPU when validating blocks. Since sync time is dominated by disk I/O, the wall-clock time for the benchmark has only a modest improvement.

benchmark: M1, SSD, sync full block

| test | before | after | after / before |
| --- | --- | --- | --- |
| `receive_block_batch()` CPU | 17.46 % | 13.57 % | 0.777 |
| wall-clock time |  109.84s | 104.46s | 0.95 |

It's also noteworthy that `Coin.name()` CPU usage goes from 3.45 % -> 0.68% a speedup of 5x.

before:
![slow-batch-00448](https://user-images.githubusercontent.com/661450/170133281-f4409b77-b51e-4b8e-8d93-cc2df7f4e8af.png)

after:
![slow-batch-00448](https://user-images.githubusercontent.com/661450/170133308-244038ce-34c8-4717-acca-61db0f411caf.png)