@arvidn 2022-05-25
I suspect this might help the wallet as well, where `Coin.name()` appears to be called a lot.