@richardkiss 2022-06-01
Wow, that's an impressive speed-up.

I agree that maybe we need to consider moving away from these fancy types. I suspect it could end up bubbling to many places